+++ CHANGELOG.md
@@ -83,6 +83,7 @@
 
 * nostr: remove `negentropy` deps ([Yuki Kishimoto])
 * nostr: remove `bitcoin` dep ([Yuki Kishimoto])
+* pool: remove `Relay::queue` ([Yuki Kishimoto])
 * connect: remove `thiserror` dep ([Yuki Kishimoto])
 * relay-builder: remove `thiserror` dep ([Yuki Kishimoto])
 * ffi: remove unnecessary `Arc` ([Yuki Kishimoto])